+++ common/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.3.0
+
+* Added optional exceptions to `GetByPath` and added tests.
+
 # v0.1.0
 
 * Initial release